% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txtheme-package.R
\docType{package}
\name{txtheme-package}
\alias{txtheme}
\alias{txtheme-package}
\title{txtheme: one palette for pkgdown, Quarto and the editor}
\description{
A colour is decided once, in one declared grid (\code{R/aaa-palette.R}), and every consumer reads what
\code{\link[=build_theme]{build_theme()}} writes from it: a pkgdown site through \code{template: package: txtheme}, a Quarto
document through the bundled extension, and the editor through a ready \code{textMateRules} block.
}
\details{
The exported functions are of two kinds: \code{\link[=build_theme]{build_theme()}}, \code{\link[=txtheme_file]{txtheme_file()}} and \code{\link[=use_brand]{use_brand()}} drive
the framework; \code{\link[=oklch_hex]{oklch_hex()}}, \code{\link[=hex_oklch]{hex_oklch()}}, \code{\link[=max_chroma]{max_chroma()}}, \code{\link[=oklch_ramp]{oklch_ramp()}}, \code{\link[=oklch_ladder]{oklch_ladder()}},
\code{\link[=apca]{apca()}} and \code{\link[=contrast]{contrast()}} are the colour maths the palette was designed with, base R only, usable
anywhere.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BriceNocenti/txtheme}
  \item Report bugs at \url{https://github.com/BriceNocenti/txtheme/issues}
}

}
\author{
\strong{Maintainer}: Brice Nocenti \email{brice.nocenti@protonmail.com}

Authors:
\itemize{
  \item Brice Nocenti \email{brice.nocenti@protonmail.com}
}

}
\keyword{internal}
